+++ spec/helpers/admin_helper_spec.rb
@@ -19,4 +19,20 @@ describe AdminHelper do
 
   end
 
+  describe '#sort_order_humanized' do
+
+    it 'returns the humanized value if one is available' do
+      expect(sort_order_humanized('name_asc')).to eq('Name ▲')
+    end
+
+    it 'returns the titleized value if a humanized value is not available' do
+      expect(sort_order_humanized('blah_asc')).to eq('Blah Asc')
+    end
+
+    it 'accepts a Symbol argument' do
+      expect(sort_order_humanized(:name_asc)).to eq('Name ▲')
+    end
+
+  end
+
 end